# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AfriLance freelance marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PortfolioItem {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  projectUrl?: string | null;
}

export interface Profile {
  id: number;
  userId: string;
  role: string;
  name: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  location?: string | null;
  skills: string[];
  /** @nullable */
  hourlyRate?: number | null;
  /** @nullable */
  category?: string | null;
  portfolioItems: PortfolioItem[];
  /** @nullable */
  avgRating?: number | null;
  totalReviews: number;
  featured: boolean;
  isBlocked: boolean;
  completedJobs: number;
  createdAt: string;
}

export interface UpsertProfileBody {
  role: string;
  name: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  location?: string | null;
  skills?: string[];
  /** @nullable */
  hourlyRate?: number | null;
  /** @nullable */
  category?: string | null;
}

export interface FreelancerListResponse {
  freelancers: Profile[];
  total: number;
}

export interface Job {
  id: number;
  clientId: number;
  title: string;
  description: string;
  category: string;
  skills: string[];
  /** @nullable */
  budgetMin?: number | null;
  /** @nullable */
  budgetMax?: number | null;
  budgetType: string;
  /** @nullable */
  location?: string | null;
  remote: boolean;
  /** @nullable */
  deadline?: string | null;
  status: string;
  isFlagged: boolean;
  proposalCount: number;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  clientAvatarUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateJobBody {
  title: string;
  description: string;
  category: string;
  skills: string[];
  /** @nullable */
  budgetMin?: number | null;
  /** @nullable */
  budgetMax?: number | null;
  budgetType: string;
  /** @nullable */
  location?: string | null;
  remote: boolean;
  /** @nullable */
  deadline?: string | null;
}

export interface UpdateJobBody {
  title?: string;
  description?: string;
  category?: string;
  skills?: string[];
  /** @nullable */
  budgetMin?: number | null;
  /** @nullable */
  budgetMax?: number | null;
  budgetType?: string;
  /** @nullable */
  location?: string | null;
  remote?: boolean;
  /** @nullable */
  deadline?: string | null;
  status?: string;
}

export interface JobListResponse {
  jobs: Job[];
  total: number;
}

export interface Proposal {
  id: number;
  jobId: number;
  freelancerId: number;
  coverLetter: string;
  bidAmount: number;
  status: string;
  /** @nullable */
  freelancerName?: string | null;
  /** @nullable */
  freelancerAvatarUrl?: string | null;
  /** @nullable */
  freelancerLocation?: string | null;
  freelancerSkills: string[];
  /** @nullable */
  freelancerAvgRating?: number | null;
  /** @nullable */
  jobTitle?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProposalBody {
  jobId: number;
  coverLetter: string;
  bidAmount: number;
}

export interface UpdateProposalStatusBody {
  status: string;
}

export interface ProposalListResponse {
  proposals: Proposal[];
  total: number;
}

export interface Conversation {
  id: number;
  otherUserId: number;
  otherUserName: string;
  /** @nullable */
  otherUserAvatarUrl?: string | null;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  lastMessageAt?: string | null;
  unreadCount: number;
}

export interface ConversationListResponse {
  conversations: Conversation[];
}

export interface Message {
  id: number;
  conversationId: number;
  senderId: number;
  content: string;
  isRead: boolean;
  createdAt: string;
}

export interface MessageListResponse {
  messages: Message[];
  total: number;
}

export interface SendMessageBody {
  recipientId: number;
  content: string;
}

export interface Notification {
  id: number;
  userId: number;
  type: string;
  title: string;
  body: string;
  isRead: boolean;
  /** @nullable */
  relatedId?: number | null;
  /** @nullable */
  relatedType?: string | null;
  createdAt: string;
}

export interface NotificationListResponse {
  notifications: Notification[];
  unreadCount: number;
}

export interface Review {
  id: number;
  jobId: number;
  reviewerId: number;
  revieweeId: number;
  rating: number;
  /** @nullable */
  comment?: string | null;
  /** @nullable */
  reviewerName?: string | null;
  /** @nullable */
  reviewerAvatarUrl?: string | null;
  createdAt: string;
}

export interface CreateReviewBody {
  jobId: number;
  revieweeId: number;
  rating: number;
  /** @nullable */
  comment?: string | null;
}

export interface ReviewListResponse {
  reviews: Review[];
  /** @nullable */
  avgRating?: number | null;
  total: number;
}

export interface FreelancerDashboard {
  totalApplications: number;
  pendingApplications: number;
  acceptedApplications: number;
  rejectedApplications: number;
  completedJobs: number;
  totalEarnings: number;
  /** @nullable */
  avgRating?: number | null;
  unreadMessages: number;
  unreadNotifications: number;
  recentProposals: Proposal[];
}

export interface ClientDashboard {
  totalJobs: number;
  openJobs: number;
  inProgressJobs: number;
  closedJobs: number;
  totalProposalsReceived: number;
  totalSpent: number;
  unreadMessages: number;
  unreadNotifications: number;
  recentJobs: Job[];
}

export interface Payment {
  id: number;
  jobId: number;
  clientId: number;
  freelancerId: number;
  amount: number;
  status: string;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  freelancerName?: string | null;
  /** @nullable */
  clientName?: string | null;
  createdAt: string;
  /** @nullable */
  releasedAt?: string | null;
}

export interface CreatePaymentBody {
  jobId: number;
  freelancerId: number;
  amount: number;
}

export interface PaymentListResponse {
  payments: Payment[];
  total: number;
}

export interface AdminUser {
  id: number;
  userId: string;
  role: string;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  location?: string | null;
  isBlocked: boolean;
  completedJobs: number;
  createdAt: string;
}

export interface AdminUserListResponse {
  users: AdminUser[];
  total: number;
}

export interface AdminBlockUserBody {
  isBlocked: boolean;
}

export interface AdminFlagBody {
  isFlagged: boolean;
}

export interface AdminStats {
  totalUsers: number;
  totalFreelancers: number;
  totalClients: number;
  totalJobs: number;
  totalProposals: number;
  totalPayments: number;
  activeJobs: number;
  flaggedJobs: number;
  blockedUsers: number;
}

export type ListFreelancersParams = {
  skill?: string;
  location?: string;
  minRate?: number;
  maxRate?: number;
  search?: string;
  featured?: boolean;
  limit?: number;
  offset?: number;
};

export type ListJobsParams = {
  category?: string;
  skill?: string;
  location?: string;
  remote?: boolean;
  minBudget?: number;
  maxBudget?: number;
  status?: string;
  search?: string;
  limit?: number;
  offset?: number;
};

export type GetConversationMessagesParams = {
  limit?: number;
  offset?: number;
};

export type ListNotificationsParams = {
  unreadOnly?: boolean;
};

export type AdminListUsersParams = {
  role?: string;
  search?: string;
  limit?: number;
  offset?: number;
};

export type AdminListJobsParams = {
  status?: string;
  flagged?: boolean;
  limit?: number;
  offset?: number;
};