# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AfriLance freelance marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List job listings with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: skill
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: remote
          in: query
          schema:
            type: boolean
        - name: minBudget
          in: query
          schema:
            type: number
        - name: maxBudget
          in: query
          schema:
            type: number
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Post a new job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobBody"
      responses:
        "200":
          description: Job updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin - list all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
  /admin/users/{id}/block:
    patch:
      operationId: adminBlockUser
      tags: [admin]
      summary: Admin - block or unblock a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBlockUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/jobs:
    get:
      operationId: adminListJobs
      tags: [admin]
      summary: Admin - list all jobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: flagged
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
  /admin/jobs/{id}/flag:
    patch:
      operationId: adminFlagJob
      tags: [admin]
      summary: Admin - flag or unflag a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminFlagBody"
      responses:
        "200":
          description: Job flagged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
components:
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        skills:
          type: array
          items:
            type: string
        budgetMin:
          type: ["number", "null"]
        budgetMax:
          type: ["number", "null"]
        budgetType:
          type: string
        location:
          type: ["string", "null"]
        remote:
          type: boolean
        deadline:
          type: ["string", "null"]
        status:
          type: string
        isFlagged:
          type: boolean
        proposalCount:
          type: integer
        clientName:
          type: ["string", "null"]
        clientAvatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - clientId
        - title
        - description
        - category
        - skills
        - budgetType
        - remote
        - status
        - isFlagged
        - proposalCount
        - createdAt
        - updatedAt
    CreateJobBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        skills:
          type: array
          items:
            type: string
        budgetMin:
          type: ["number", "null"]
        budgetMax:
          type: ["number", "null"]
        budgetType:
          type: string
        location:
          type: ["string", "null"]
        remote:
          type: boolean
        deadline:
          type: ["string", "null"]
      required:
        - title
        - description
        - category
        - skills
        - budgetType
        - remote
    UpdateJobBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        skills:
          type: array
          items:
            type: string
        budgetMin:
          type: ["number", "null"]
        budgetMax:
          type: ["number", "null"]
        budgetType:
          type: string
        location:
          type: ["string", "null"]
        remote:
          type: boolean
        deadline:
          type: ["string", "null"]
        status:
          type: string
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
      required:
        - jobs
        - total
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        role:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        isBlocked:
          type: boolean
        completedJobs:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - role
        - name
        - isBlocked
        - completedJobs
        - createdAt
    AdminUserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
      required:
        - users
        - total
    AdminBlockUserBody:
      type: object
      properties:
        isBlocked:
          type: boolean
      required:
        - isBlocked
    AdminFlagBody:
      type: object
      properties:
        isFlagged:
          type: boolean
      required:
        - isFlagged
...