# lib/api-zod/src/generated/types/updateJobBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AfriLance freelance marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateJobBody {
  title?: string;
  description?: string;
  category?: string;
  skills?: string[];
  /** @nullable */
  budgetMin?: number | null;
  /** @nullable */
  budgetMax?: number | null;
  budgetType?: string;
  /** @nullable */
  location?: string | null;
  remote?: boolean;
  /** @nullable */
  deadline?: string | null;
  status?: string;
}